Moved usershelper to helpers from middlewares
